Order fixed and reimplemented

+++ controllers/Order.js
@@ -1,52 +1,8 @@
 'use strict';
 
-var utils = require('../utils/writer.js');
-var Order = require('../service/OrderService');
-var session = require('../utils/SessionManager');
-
-
-//TODO IMPLEMENTARE TUTTO
-
-module.exports.orderAddressGET = function orderAddressGET (req, res, next) {
-    var orderId = req.swagger.params['orderId'].value;
-
-    if(!session.userIdExists()){
-        utils.writeJson(res, {message: "You must login to perform this operation!"}, 403);
-        return;
-    }
-
-    let userId = session.getUserId();
-
-    Order.orderAddressGET(orderId, userId)
-        .then(function (response) {
-            utils.writeJson(res, response);
-        })
-        .catch(function (response) {
-            utils.writeJson(res, response);
-        });
-};
-
-module.exports.orderDetailsGET = function orderDetailsGET (req, res, next) {
-
-    if(!session.userIdExists()){
-        utils.writeJson(res, {message: "You must login to perform this operation!"}, 403);
-        return;
-    }
-
-    let userId = session.getUserId();
-
-
-    var orderId = req.swagger.params['orderId'].value;
-
-
-    Order.orderDetailsGET(orderId, userId)
-        .then(function (response) {
-            utils.writeJson(res, response);
-        })
-        .catch(function (response) {
-            utils.writeJson(res, response, response && response.errorCode || 500);
-        });
-};
+const utils = require('../utils/writer.js'),
+      Order = require('../service/OrderService'),
+      session = require('../utils/SessionManager');
 
 module.exports.orderPlacePOST = function orderPlacePOST (req, res, next) {
     var addressStreetLine1 = req.swagger.params['addressStreetLine1'].value;
@@ -54,9 +10,11 @@ module.exports.orderPlacePOST = function orderPlacePOST (req, res, next) {
     var zip_code = req.swagger.params['zip_code'].value;
     var province = req.swagger.params['province'].value;
     var country = req.swagger.params['country'].value;
-    var addressStreetLine2 = req.swagger.params['addressStreetLine2'].value;
+    var shipping_method = req.swagger.params['shipping_method'].value;
+    var payment_method = req.swagger.params['payment_method'].value;
     var firstName = req.swagger.params['firstName'].value;
     var lastName = req.swagger.params['lastName'].value;
+    var addressStreetLine2 = req.swagger.params['addressStreetLine2'].value;
 
     if(!session.userIdExists()){
         utils.writeJson(res, {message: "You must login to perform this operation!"}, 403);
@@ -65,21 +23,39 @@ module.exports.orderPlacePOST = function orderPlacePOST (req, res, next) {
 
     let userId = session.getUserId();
 
-    Order.orderPlacePOST(userId,addressStreetLine1,city,zip_code,province,country,firstName,lastName,addressStreetLine2)
+    Order.orderPlacePOST(userId,addressStreetLine1,city,zip_code,province,country,shipping_method,payment_method,firstName,lastName,addressStreetLine2)
         .then(function (response) {
-            utils.writeJson(res, response);
+            console.log(response);
+            utils.writeJson(res,
+                            {orderId: response[0]});
         })
         .catch(function (response) {
-            utils.writeJson(res, response);
+            console.error(response);
+            utils.writeJson(res,
+                            {message: response.message || "Internal Server Error!"},
+                            response.errorCode || 500);
         });
 };
 
 module.exports.ordersGET = function ordersGET (req, res, next) {
     var offset = req.swagger.params['offset'].value;
     var limit = req.swagger.params['limit'].value;
-    Order.ordersGET(offset,limit)
+
+    if(!session.userIdExists()){
+        utils.writeJson(res, {message: "You must login to perform this operation!"}, 403);
+        return;
+    }
+
+    let userId = session.getUserId();
+
+    Order.ordersGET(userId,offset,limit)
         .then(function (response) {
-            utils.writeJson(res, response);
+            if(response.length)
+                utils.writeJson(res, response);
+            else
+                utils.writeJson(res,
+                                {message: "No orders found!"},
+                                404);
         })
         .catch(function (response) {
             utils.writeJson(res, response);
